+++ changelog.md
@@ -2,8 +2,9 @@
 
 ## v0.6.0
 
-*  Ability to toggle to read with screen reader in chat menu.
+*  Windows only: Ability to toggle to read with screen reader in chat menu.
 * Reset settings in advance.
+* Warns to reset the settings not compatible with the current version.
 
 ## v0.5.0